+++ packages/tokenizer/src/grammar-loader.ts
@@ -144,6 +144,7 @@ import * as yaml from 'reprism/languages/yaml';
 import * as tap from './languages/tap';
 import * as sql from './languages/sql';
 import * as plsql from './languages/plsql';
+import * as gdscript from './languages/gdscript';
 
 export const languages = {
   abap, actionscript, ada, apacheconf, apl, applescript, arff,
@@ -159,7 +160,7 @@ export const languages = {
   properties, protobuf, pug, puppet, pure, python, q, qore, r, reason, renpy, rest,
   rip, roboconf, ruby, rust, sas, sass, scala, scheme, scss, smalltalk, smarty, soy,
   stylus, swift, tcl, textile, twig, typescript, vbnet, velocity, verilog, vhdl,
-  vim, visualBasic, wasm, wiki, xeora, xojo, yaml, tsx, sql, plsql, tap
+  vim, visualBasic, wasm, wiki, xeora, xojo, yaml, tsx, sql, plsql, tap, gdscript
 };
 
 export const loadLanguages = (): void => {