+++ apps/jscpd/CHANGELOG.md
@@ -1,5 +1,31 @@
 ## 4.0.0
 
+## 4.0.7
+
+### Patch Changes
+
+- fix issue with build
+- Updated dependencies
+  - @jscpd/badge-reporter@4.0.3
+  - jscpd-sarif-reporter@4.0.5
+  - @jscpd/html-reporter@4.0.3
+  - @jscpd/tokenizer@4.0.3
+  - @jscpd/finder@4.0.3
+  - @jscpd/core@4.0.3
+
+## 4.0.6
+
+### Patch Changes
+
+- small fixes, merge suggested PRs on github
+- Updated dependencies
+  - @jscpd/badge-reporter@4.0.2
+  - @jscpd/core@4.0.2
+  - @jscpd/finder@4.0.2
+  - @jscpd/html-reporter@4.0.2
+  - jscpd-sarif-reporter@4.0.4
+  - @jscpd/tokenizer@4.0.2
+
 ## 4.0.4
 
 ### Patch Changes