+++ packages/html-reporter/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @jscpd/html-reporter
 
+## 4.0.3
+
+### Patch Changes
+
+- fix issue with build
+
+## 4.0.2
+
+### Patch Changes
+
+- small fixes, merge suggested PRs on github
+
 ## 4.0.1
 
 ### Patch Changes