+++ packages/redis-store/CHANGELOG.md
@@ -1,5 +1,23 @@
 ## 3.3.0-rc.3 (2020-05-04)
 
+## 4.0.3
+
+### Patch Changes
+
+- fix issue with build
+- Updated dependencies
+  - @jscpd/tokenizer@4.0.3
+  - @jscpd/core@4.0.3
+
+## 4.0.2
+
+### Patch Changes
+
+- small fixes, merge suggested PRs on github
+- Updated dependencies
+  - @jscpd/core@4.0.2
+  - @jscpd/tokenizer@4.0.2
+
 ## 4.0.1
 
 ### Patch Changes