+++ packages/tsconfig/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @jscpd/tsconfig
 
+## 5.0.3
+
+### Patch Changes
+
+- fix issue with build
+
+## 5.0.2
+
+### Patch Changes
+
+- small fixes, merge suggested PRs on github
+
 ## 5.0.1
 
 ### Patch Changes